change second from int to double

+++ collection/util/time/Time.cpp
@@ -6,6 +6,6 @@ uint32_t Time::getMs() {
   return SDL_GetTicks();
 }
 
-uint32_t Time::getSeconds() {
-  return SDL_GetTicks() / 1000;
+double Time::getSeconds() {
+  return static_cast<double>(SDL_GetTicks()) / 1000.0;
 }
