+++ collection/util/time/Time.h
@@ -6,5 +6,5 @@ class Time {
  public:
   /* Get Ms since starting */
   static uint32_t getMs();
-  static uint32_t getSeconds();
+  static double getSeconds();
 };
